requirements.txt
anyio==4.6.2.post1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.8.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
cloud-init==23.2.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
decorator==4.4.2
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
Flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
lark-oapi==1.3.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.1
pexpect==4.8.0
pillow==11.0.0
protobuf==3.20.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycryptodome==3.21.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytesseract==0.3.13
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-linux-procfs==0.6.3
python-magic==0.4.24
pytz==2022.1
pyudev==0.22.0
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==1.0.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sos==4.5.6
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
websockets==13.1
Werkzeug==3.0.4
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
